Move route models to separate directory
